create special_operators module and reorganize
