+++ src/openfermion/hamiltonians/_mean_field_dwave.py
@@ -13,10 +13,9 @@
 """This module constructs Hamiltonians for the BCS mean-field d-wave model."""
 from __future__ import absolute_import
 
-from openfermion.hamiltonians import up_index, down_index
+from openfermion.hamiltonians import up_index, down_index, number_operator
 from openfermion.ops import (FermionOperator,
-                             hermitian_conjugated,
-                             number_operator)
+                             hermitian_conjugated)
 
 
 def mean_field_dwave(x_dimension, y_dimension, tunneling, sc_gap,
