+++ src/openfermion/hamiltonians/_special_operators_test.py
