+++ src/openfermion/ops/__init__.py
@@ -14,9 +14,7 @@ from ._binary_operator import SymbolicBinary
 from ._symbolic_operator import SymbolicOperator
 from ._fermion_operator import (FermionOperator,
                                 hermitian_conjugated,
-                                majorana_operator,
-                                normal_ordered,
-                                number_operator)
+                                normal_ordered)
 from ._qubit_operator import QubitOperator
 from ._polynomial_tensor import general_basis_change, PolynomialTensor
 from ._interaction_operator import InteractionOperator
