+++ src/openfermion/ops/_fermion_operator.py
@@ -31,74 +31,6 @@ def hermitian_conjugated(fermion_operator):
     return conjugate_operator
 
 
-def majorana_operator(term=None, coefficient=1.):
-    """Initialize a Majorana operator.
-
-    Args:
-        term(tuple): The first element of the tuple indicates the mode
-            on which the Majorana operator acts, starting from zero.
-            The second element of the tuple is an integer, either 1 or 0,
-            indicating which type of Majorana operator it is:
-
-                Type 1: :math:`\\frac{1}{\sqrt{2}} (a^\dagger_p + a_p)`
-
-                Type 0: :math:`\\frac{i}{\sqrt{2}} (a^\dagger_p - a_p)`
-
-            where the :math:`a^\dagger_p` and :math:`a_p` are the usual
-            fermionic ladder operators.
-            Default will result in the zero operator.
-        coefficient(complex or float, optional): The coefficient of the term.
-            Default value is 1.0.
-
-    Returns:
-        FermionOperator
-    """
-    if not isinstance(coefficient, (int, float, complex)):
-        raise ValueError('Coefficient must be scalar.')
-
-    if term is None:
-        # Return zero operator
-        return FermionOperator()
-    elif isinstance(term, tuple):
-        mode, operator_type = term
-        if operator_type == 1:
-            majorana_op = FermionOperator(
-                ((mode, 1),), coefficient / numpy.sqrt(2.))
-            majorana_op += FermionOperator(
-                ((mode, 0),), coefficient / numpy.sqrt(2.))
-        elif operator_type == 0:
-            majorana_op = FermionOperator(
-                ((mode, 1),), 1.j * coefficient / numpy.sqrt(2.))
-            majorana_op -= FermionOperator(
-                ((mode, 0),), 1.j * coefficient / numpy.sqrt(2.))
-        else:
-            raise ValueError('Operator specified incorrectly.')
-        return majorana_op
-    else:
-        # Invalid input.
-        raise ValueError('Operator specified incorrectly.')
-
-
-def number_operator(n_orbitals, orbital=None, coefficient=1.):
-    """Return a number operator.
-
-    Args:
-        n_orbitals (int): The number of spin-orbitals in the system.
-        orbital (int, optional): The orbital on which to return the number
-            operator. If None, return total number operator on all sites.
-        coefficient (float): The coefficient of the term.
-    Returns:
-        operator (FermionOperator)
-    """
-    if orbital is None:
-        operator = FermionOperator()
-        for spin_orbital in range(n_orbitals):
-            operator += number_operator(n_orbitals, spin_orbital, coefficient)
-    else:
-        operator = FermionOperator(((orbital, 1), (orbital, 0)), coefficient)
-    return operator
-
-
 def normal_ordered_term(term, coefficient):
     """Return a normal ordered FermionOperator corresponding to single term.
