+++ src/openfermion/transforms/_conversion_test.py
@@ -17,8 +17,9 @@ import copy
 import numpy
 import unittest
 
+from openfermion.hamiltonians import number_operator
 from openfermion.ops import (FermionOperator, InteractionOperator,
-                             normal_ordered, number_operator, QubitOperator)
+                             normal_ordered, QubitOperator)
 from openfermion.ops._interaction_operator import InteractionOperatorError
 from openfermion.ops._quadratic_hamiltonian import QuadraticHamiltonianError
 from openfermion.transforms import *
